+++ contracts/interfaces/IBroker.sol
@@ -2,6 +2,7 @@
 pragma solidity 0.8.19;
 
 import "@openzeppelin/contracts/token/ERC20/extensions/IERC20Metadata.sol";
+import "@reserve-protocol/trusted-fillers/contracts/interfaces/ITrustedFillerRegistry.sol";
 import "./IAsset.sol";
 import "./IComponent.sol";
 import "./IGnosis.sol";
@@ -40,6 +41,7 @@ interface IBroker is IComponent {
     event DutchAuctionLengthSet(uint48 oldVal, uint48 newVal);
     event BatchTradeDisabledSet(bool prevVal, bool newVal);
     event DutchTradeDisabledSet(IERC20Metadata indexed erc20, bool prevVal, bool newVal);
+    event TrustedFillerRegistrySet(address trustedFillerRegistry, bool isEnabled);
 
     // Initialization
     function init(
@@ -71,6 +73,12 @@ interface IExtendedBroker is IBroker {
     function setBatchTradeImplementation(ITrade newTradeImplementation) external;
 
     function setDutchTradeImplementation(ITrade newTradeImplementation) external;
+
+    function setTrustedFillerRegistry(address newRegistry, bool enabled) external;
+
+    function trustedFillerRegistry() external view returns (ITrustedFillerRegistry);
+
+    function trustedFillerEnabled() external view returns (bool);
 }
 
 interface TestIBroker is IBroker {
@@ -98,4 +106,10 @@ interface TestIBroker is IBroker {
 
     // only present on pre-3.0.0 Brokers; used by EasyAuction regression test
     function disabled() external view returns (bool);
+
+    function setTrustedFillerRegistry(address newRegistry, bool enabled) external;
+
+    function trustedFillerRegistry() external view returns (ITrustedFillerRegistry);
+
+    function trustedFillerEnabled() external view returns (bool);
 }