+++ test/plugins/individual-collateral/compoundv2/CTokenFiatCollateral.test.ts
@@ -268,6 +268,7 @@ describeFork(`CTokenFiatCollateral - Mainnet Forking P${IMPLEMENTATION}`, functi
         assetPluginRegistry: ZERO_ADDRESS,
         daoFeeRegistry: ZERO_ADDRESS,
         versionRegistry: ZERO_ADDRESS,
+        trustedFillerRegistry: ZERO_ADDRESS,
       })
     ).wait()